+++ src/LibraryADT.java
@@ -1,12 +1,12 @@
 public interface LibraryADT {
     boolean logIn(String userID, String role);
-    boolean addBook(int isbn, String title, String author);
-    boolean removeBook(int isbn);
-    void searchByIsbn(int isbn);
+    boolean addBook(long isbn, String title, String author);
+    boolean removeBook(long isbn);
+    void searchByIsbn(long isbn);
     void searchBooksByTitle(String title);
     void searchBooksByAuthor(String author);
-    boolean borrowBook(int isbn, String userID);
-    boolean returnBook(int isbn, String userID);
+    boolean borrowBook(long isbn, String userID);
+    boolean returnBook(long isbn, String userID);
     void viewAllBooks();
     void viewAllBorrowedBooks();
     void viewOwnBorrowedBooks(String userID);
